Create an abstract super-class for View controllers.
