+++ src/controller/menus/mainMenu/MainMenuController.java
@@ -10,9 +10,9 @@
 
 package controller.menus.mainMenu;
 
+import controller.AbstractViewController;
 import controller.DialogBoxFactory;
 import controller.ViewController;
-import controller.ViewLoader;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ButtonType;
@@ -20,36 +20,34 @@ import javafx.stage.Stage;
 import javafx.stage.WindowEvent;
 
 import java.util.Optional;
-import java.util.ResourceBundle;
 
 /**
  * Contrôleur associé au menu principal.
  * <p>
- * Le menu est principalement composé de quatre boutons :
- *      - Un bouton "Nouvelle Partie", lançant une partie <i>from scratch</i>. [non-implémenté]
- *      - Un bouton "Continuer une Partie", lançant une partie depuis une sauvegarde. [non-implémenté]
- *      - Un bouton "Options", pour changer quelques paramètres. [non-implémentés]
+ * Le menu est principalement composé de quatre boutons : <br>
+ *      - Un bouton "Nouvelle Partie", lançant une partie <i>from scratch</i>. [non-implémenté] <br>
+ *      - Un bouton "Continuer une Partie", lançant une partie depuis une sauvegarde. [non-implémenté] <br>
+ *      - Un bouton "Options", pour changer quelques paramètres. [non-implémentés] <br>
  *      - Un bouton "Quitter", lançant une requête pour fermer l'application.
+ *
  * <p>
- * Comporte également un attribut pour le {@link Stage} afin de lancer soi-même des requêtes de fermetures,
- * ainsi qu'une {@link DialogBoxFactory fabrique à boîte de dialogue}, pour simplifier la lisibilité de la classe, et
+ * Utilise les instances de la {@link AbstractViewController classe-mère} :
+ *      - Le {@link Stage} afin de lancer soi-même des requêtes de fermetures. <br>
+ *      - La {@link DialogBoxFactory fabrique à boîte de dialogue}, pour simplifier la lisibilité de la classe, et
  * séparer les responsabilités selon un <i>design pattern</i> de fabrique.
- * <p>
- * Ces attributs exigent des éléments créés par les classes chargeant le FXML, qui doivent donc être transférés,
- * au moyen des méthodes {@link #initStage(Stage)} et {@link #initBundle(ResourceBundle)} respectivement.
- * <p>
+ *
  * <p>
  * Les attributs de classe <i>FXMLPath</i> et <i>MainMenuCSS</i> ne changent pas d'un lancement à un autre,
- * et sont utilisés par plusieurs classes ({@link application.Main} et le future <i>GameController</i>.
+ * et sont utilisés par plusieurs classes ({@link application.Main}, les sous-menus et le future <i>GameController</i>.
+ *
+ * @see AbstractViewController
+ * @see ViewController
  */
-public class MainMenuController implements ViewController {
+public class MainMenuController extends AbstractViewController implements ViewController {
     /** Chemin menant au fichier FXML associé au menu principal. */
     public static final String FXMLPath = "/controller/menus/mainMenu/MainMenu.fxml";
     /** Chemin menant au fichier CSS associé au menu, <b>sans</b> le masque de thème et l'extension. */
     public static final String CSSPath = "controller/menus/mainMenu/MainMenu";
-
-    /** Fabrique à boîte de dialogue associée à ce contrôleur. */
-    private DialogBoxFactory dialogFactory;
 
     /** Bouton "Nouvelle Partie". */
     @FXML private Button newGameButton;
@@ -57,9 +55,6 @@ public class MainMenuController implements ViewController {
     @FXML private Button continueGameButton;
     /** Bouton "Options". */
     @FXML private Button settingsButton;
-
-    /** Stage associé à ce contrôleur. */
-    private Stage stage;
 
     /**
      * Actions effectuées durant le {@link javafx.fxml.FXMLLoader chargement de la vue FXML}.
@@ -75,73 +70,6 @@ public class MainMenuController implements ViewController {
         this.newGameButton.setDisable(true);
         this.continueGameButton.setDisable(true);
         this.settingsButton.setDisable(true);
-    }
-
-    /**
-     * Transfère l'attribut {@link Stage} depuis l'instance {@link ViewLoader} accèdant au menu principal.
-     *
-     * <p>
-     * Modifie la propriété <i>onCloseRequest</i> du {@link Stage} afin d'appeler {@link
-     * #isCloseable()} et de devoir obtenir une validation avant de fermer l'application.
-     *
-     * @param _stage
-     *         Le {@link Stage stage} associé à ce contrôleur.
-     *
-     * @throws IllegalStateException
-     *         Lancée si le {@link Stage} existe déjà à l'appel de cette méthode. <br>
-     *         N'est pas censée être rattrapée, étant un problème de programmation.
-     */
-    @Override
-    public void initStage(Stage _stage) {
-        if (this.stage != null) {
-            throw new IllegalStateException("The stage already exists");
-        }
-        this.stage = _stage;
-
-        this.stage.setOnCloseRequest(event -> {
-            if (this.isCloseable()) {
-                this.stage.close();
-            } else {
-                event.consume();
-            }
-        });
-    }
-
-    /**
-     * Transfère une instance {@link ResourceBundle} vers l'instance {@link DialogBoxFactory} afin d'obtenir des
-     * boîtes de dialogues changeantes selon la langue de l'application.
-     * <p>
-     * Devrait être appelée par l'instance {@link ViewLoader} qui a instancié le pack.
-     *
-     * @param _bundle
-     *         Le {@link ResourceBundle pack de ressources} associé à ce contrôleur.
-     *
-     * @throws IllegalArgumentException
-     *         Lancée si le {@link ResourceBundle pack} existe déjà à l'appel de cette méthode et ne change pas de
-     *         langue. <br>
-     *         N'est pas censée être rattrapée, étant un problème de programmation.
-     * @see DialogBoxFactory
-     */
-    @Override
-    public void initBundle(ResourceBundle _bundle) {
-        if (dialogFactory == null) {
-            this.dialogFactory = new DialogBoxFactory(_bundle);
-        } else {
-            this.dialogFactory.setBundle(_bundle);
-        }
-    }
-
-    /**
-     * Permet l'accès au chemin du fichier CSS à travers une instance.
-     * <p>
-     * Cela permet à {@link ViewLoader} de ne pas avoir à utiliser un paramètre supplémentaire dans le constructeur,
-     * ainsi de ne pas forcer l'attribut <i>CSSPath</i> en tant qu'attribut de classe ou d'instance.
-     *
-     * @return Retourne le chemin vers le fichier CSS associé au contrôleur.
-     */
-    @Override
-    public String getCSSPath() {
-        return MainMenuController.CSSPath;
     }
 
     /**
@@ -194,8 +122,8 @@ public class MainMenuController implements ViewController {
      * @see DialogBoxFactory
      * @see javafx.scene.control.Alert
      */
-    @FXML
-    private boolean isCloseable() {
+    @Override
+    protected boolean isCloseable() {
         Optional<ButtonType> result = this.dialogFactory.CloseRequestDialog().showAndWait();
         return result.isPresent() && result.get().equals(ButtonType.OK);
     }
